Sign in to Jenkins with valid/invalid cred
